migrate and update schema
Note: there's several difs between my generated schema and the
previous version that aren't caused by the migration
(and so aren't reflected in this commit.) The comments are different,
the strings are "double quoted", and the elections.settings switch to
jsonb isn't generated (as it appears to be a manual change).

Ideally, we'd update schema.json to reflect the trivial diffs so that
we can regenerate schema.rb without a lot of diff noise.
elections.settings would be tricker to accomplish.