Record the current user for the session/thread
We want to record who made the change to records via a commit hook.
The commit hook doesn't have access to the controller (and with it
Devise's current_user), so we'll set that manually on the
current Thread. This way, everything on the thread (including model
commit hooks) will have easy access to the user, thus enabling auditing.